GAN/requirements.txt
pytorch>=1.1.0
torchvision>=0.3.0
tensorboard>=1.12.2
tensorboardX>=1.6
tensorflow-gpu==1.12.0
